refactor: direct parse packet msg for udp
To optimize udp service throughput
