+++ sip/parser/common.go
@@ -1,17 +1,12 @@
-// Forked from github.com/StefanKopieczek/gossip by @StefanKopieczek
 package parser
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"strconv"
 	"strings"
-	"sync"
 	"unicode"
 	"unicode/utf8"
-
-	"github.com/tevino/abool"
 
 	"github.com/ghettovoice/gosip/log"
 	"github.com/ghettovoice/gosip/sip"
@@ -25,17 +20,15 @@ const abnfWs = " \t"
 // C.f. RFC 3261 S. 8.1.1.5.
 const maxCseq = 2147483647
 
-// The buffer size of the parser input channel.
-
 // A Parser converts the raw bytes of SIP messages into core.Message objects.
 // It allows
 type Parser interface {
-	// Implements io.Writer. Queues the given bytes to be parsed.
+	// Write implements io.Writer. Queues the given bytes to be parsed.
 	// If the parser has terminated due to a previous fatal error, it will return n=0 and an appropriate error.
 	// Otherwise, it will return n=len(p) and err=nil.
 	// Note that err=nil does not indicate that the data provided is valid - simply that the data was successfully queued for parsing.
 	Write(p []byte) (n int, err error)
-	// Register a custom header parser for a particular header type.
+	// SetHeaderParser register a custom header parser for a particular header type.
 	// This will overwrite any existing registered parser for that header type.
 	// If a parser is not available for a header type in a message, the parser will produce a core.GenericHeader struct.
 	SetHeaderParser(headerName string, headerParser HeaderParser)
@@ -86,433 +79,10 @@ func defaultHeaderParsers() map[string]HeaderParser {
 	}
 }
 
-// Parse a SIP message by creating a parser on the fly.
-// This is more costly than reusing a parser, but is necessary when we do not
-// have a guarantee that all messages coming over a connection are from the
-// same endpoint (e.g. UDP).
+// ParseMessage parse a SIP message by creating a parser on the fly.
 func ParseMessage(msgData []byte, logger log.Logger) (sip.Message, error) {
-	output := make(chan sip.Message, 0)
-	errs := make(chan error, 0)
-
-	parser := NewParser(output, errs, false, logger)
-	defer parser.Stop()
-
-	if _, err := parser.Write(msgData); err != nil {
-		return nil, err
-	}
-
-	select {
-	case msg := <-output:
-		return msg, nil
-	case err := <-errs:
-		return nil, err
-	}
-}
-
-type PacketParser struct {
-	p    Parser
-	out  chan sip.Message
-	errs chan error
-}
-
-func NewPacketParser(logger log.Logger) *PacketParser {
-	output := make(chan sip.Message, 0)
-	errs := make(chan error, 0)
-
-	return &PacketParser{
-		p:    NewParser(output, errs, false, logger),
-		out:  output,
-		errs: errs,
-	}
-}
-
-func (pp *PacketParser) ParseMessage(msgData []byte) (sip.Message, error) {
-	//defer pp.p.Reset()
-
-	if _, err := pp.p.Write(msgData); err != nil {
-		return nil, err
-	}
-
-	select {
-	case msg := <-pp.out:
-		return msg, nil
-	case err := <-pp.errs:
-		return nil, err
-	}
-}
-
-func (pp *PacketParser) Stop() {
-	if pp == nil {
-		return
-	}
-	if pp.p != nil {
-		pp.p.Stop()
-	}
-	close(pp.out)
-	close(pp.errs)
-}
-
-// Create a new Parser.
-//
-// Parsed SIP messages will be sent down the 'output' chan provided.
-// Any errors which force the parser to terminate will be sent down the 'errs' chan provided.
-//
-// If streamed=false, each Write call to the parser should contain data for one complete SIP message.
-
-// If streamed=true, Write calls can contain a portion of a full SIP message.
-// The end of one message and the start of the next may be provided in a single call to Write.
-// When streamed=true, all SIP messages provided must have a Content-Length header.
-// SIP messages without a Content-Length will cause the parser to permanently stop, and will result in an error on the errs chan.
-
-// 'streamed' should be set to true whenever the caller cannot reliably identify the starts and ends of messages from the transport frames,
-// e.g. when using streamed protocols such as TCP.
-func NewParser(
-	output chan<- sip.Message,
-	errs chan<- error,
-	streamed bool,
-	logger log.Logger,
-) Parser {
-	p := &parser{
-		streamed: streamed,
-		done:     make(chan struct{}),
-	}
-	p.log = logger.
-		WithPrefix("parser.Parser").
-		WithFields(log.Fields{
-			"parser_ptr": fmt.Sprintf("%p", p),
-		})
-	// Configure the parser with the standard set of header parsers.
-	p.headerParsers = make(map[string]HeaderParser)
-	for headerName, headerParser := range defaultHeaderParsers() {
-		p.SetHeaderParser(headerName, headerParser)
-	}
-
-	p.output = output
-	p.errs = errs
-
-	// Create a managed buffer to allow message data to be asynchronously provided to the parser, and
-	// to allow the parser to block until enough data is available to parse.
-	p.input = newParserBuffer(p.Log())
-	// Done for input a line at a time, and produce SipMessages to send down p.output.
-	go p.parse(streamed, p.done)
-
-	return p
-}
-
-type parser struct {
-	headerParsers map[string]HeaderParser
-	streamed      bool
-	input         *parserBuffer
-
-	output chan<- sip.Message
-	errs   chan<- error
-
-	stopped abool.AtomicBool
-
-	mu   sync.Mutex
-	done chan struct{}
-
-	log log.Logger
-}
-
-func (p *parser) String() string {
-	if p == nil {
-		return "Parser <nil>"
-	}
-	return fmt.Sprintf("Parser %p", p)
-}
-
-func (p *parser) Log() log.Logger {
-	return p.log
-}
-
-func (p *parser) Write(data []byte) (int, error) {
-	if p.stopped.IsSet() {
-		return 0, WriteError(fmt.Sprintf("cannot write data to stopped %s", p))
-	}
-
-	var (
-		num int
-		err error
-	)
-	if !p.streamed {
-		bl := getBodyLength(data)
-
-		if bl == -1 {
-			err = InvalidMessageFormat(fmt.Sprintf("%s cannot write data: double CRLF sequence not found in the input data", p))
-			return num, err
-		}
-
-		data = append([]byte(fmt.Sprintf("%d|%d\r\n", bl, len(data))), data...)
-	}
-
-	num, err = p.input.Write(data)
-	if err != nil {
-		err = WriteError(fmt.Sprintf("%s write data failed: %s", p, err))
-		return num, err
-	}
-
-	p.Log().Tracef("write %d bytes to input buffer", num)
-
-	return num, nil
-}
-
-// Stop parser processing, and allow all resources to be garbage collected.
-// The parser will not release its resources until Stop() is called,
-// even if the parser object itself is garbage collected.
-func (p *parser) Stop() {
-	if p.stopped.IsSet() {
-		return
-	}
-
-	p.Log().Debug("stopping parser...")
-
-	p.stopped.Set()
-
-	p.input.Stop()
-	p.mu.Lock()
-	done := p.done
-	p.mu.Unlock()
-	<-done
-
-	p.Log().Debug("parser stopped")
-}
-
-func (p *parser) Reset() {
-	p.Stop()
-	// reset state
-	p.mu.Lock()
-	done := make(chan struct{})
-	p.done = done
-	p.mu.Unlock()
-	p.input.Reset()
-	// and re-run
-	go p.parse(p.streamed, done)
-
-	p.stopped.UnSet()
-}
-
-// Consume input lines one at a time, producing core.Message objects and sending them down p.output.
-func (p *parser) parse(requireContentLength bool, done chan<- struct{}) {
-	defer close(done)
-
-	var msg sip.Message
-
-	p.Log().Debug("start parsing")
-	defer p.Log().Debug("stop parsing")
-
-	var skipStreamedErr bool
-
-	for {
-		var bodyLen, msgLen int
-		if !p.streamed {
-			// extract body/msg len
-			line, err := p.input.NextLine()
-			if err != nil {
-				break
-			}
-			strs := strings.Split(line, "|")
-			if len(strs) != 2 {
-				continue
-			}
-			bodyLen, err = strconv.Atoi(strs[0])
-			if err != nil {
-				continue
-			}
-			msgLen, err = strconv.Atoi(strs[1])
-			if err != nil {
-				continue
-			}
-		}
-		// Parse the StartLine.
-		startLine, err := p.input.NextLine()
-		if err != nil {
-			break
-		}
-
-		p.Log().Tracef("start reading start line: %s", startLine)
-
-		var termErr error
-		if isRequest(startLine) {
-			method, recipient, sipVersion, err := ParseRequestLine(startLine)
-			if err == nil {
-				msg = sip.NewRequest("", method, recipient, sipVersion, []sip.Header{}, "", nil)
-			} else {
-				termErr = err
-			}
-		} else if isResponse(startLine) {
-			sipVersion, statusCode, reason, err := ParseStatusLine(startLine)
-			if err == nil {
-				msg = sip.NewResponse("", sipVersion, statusCode, reason, []sip.Header{}, "", nil)
-			} else {
-				termErr = err
-			}
-		} else {
-			termErr = fmt.Errorf("transmission beginning '%s' is not a SIP message", startLine)
-		}
-
-		if termErr != nil {
-			p.Log().Tracef("%s failed to read start line '%s'", p, startLine)
-
-			termErr = InvalidStartLineError(fmt.Sprintf("%s failed to parse first line of message: %s", p, termErr))
-
-			if p.streamed {
-				if !skipStreamedErr {
-					skipStreamedErr = true
-					p.errs <- termErr
-				}
-			} else {
-				skip := msgLen - len(startLine) - 2
-
-				p.Log().Tracef("skip %d - %d - 2 = %d bytes", p, msgLen, len(startLine), skip)
-
-				if _, err := p.input.NextChunk(skip); err != nil {
-					p.Log().Errorf("skip failed: %s", err)
-				}
-
-				p.errs <- termErr
-			}
-
-			continue
-		} else {
-			skipStreamedErr = false
-		}
-
-		p.Log().Tracef("%s starts reading headers", p)
-		// Parse the header section.
-		// Headers can be split across lines (marked by whitespace at the start of subsequent lines),
-		// so store lines into a buffer, and then flush and parse it when we hit the end of the header.
-		var buffer bytes.Buffer
-		headers := make([]sip.Header, 0)
-
-		flushBuffer := func() {
-			if buffer.Len() > 0 {
-				newHeaders, err := p.ParseHeader(buffer.String())
-				if err == nil {
-					headers = append(headers, newHeaders...)
-				} else {
-					p.Log().Warnf("skip header '%s' due to error: %s", buffer, err)
-				}
-				buffer.Reset()
-			}
-		}
-
-		for {
-			line, err := p.input.NextLine()
-
-			if err != nil {
-				break
-			}
-
-			if len(line) == 0 {
-				// We've hit the end of the header section.
-				// Parse anything remaining in the buffer, then break out.
-				flushBuffer()
-
-				break
-			}
-
-			if !strings.Contains(abnfWs, string(line[0])) {
-				// This line starts a new header.
-				// Parse anything currently in the buffer, then store the new header line in the buffer.
-				flushBuffer()
-				buffer.WriteString(line)
-			} else if buffer.Len() > 0 {
-				// This is a continuation line, so just add it to the buffer.
-				buffer.WriteString(" ")
-				buffer.WriteString(line)
-			} else {
-				// This is a continuation line, but also the first line of the whole header section.
-				// Discard it and log.
-				p.Log().Tracef(
-					"discard unexpected continuation line '%s' at start of header block in message '%s'",
-					line,
-					msg.Short(),
-				)
-			}
-		}
-
-		// Store the headers in the message object.
-		for _, header := range headers {
-			msg.AppendHeader(header)
-		}
-
-		var contentLength int
-		// Determine the length of the body, so we know when to stop parsing this message.
-		if p.streamed {
-			// Use the content-length header to identify the end of the message.
-			contentLengthHeaders := msg.GetHeaders("Content-Length")
-			if len(contentLengthHeaders) == 0 {
-				skipStreamedErr = true
-
-				termErr := &sip.MalformedMessageError{
-					Err: fmt.Errorf("missing required 'Content-Length' header"),
-					Msg: msg.String(),
-				}
-				p.errs <- termErr
-
-				continue
-			} else if len(contentLengthHeaders) > 1 {
-				skipStreamedErr = true
-
-				var errbuf bytes.Buffer
-				errbuf.WriteString("multiple 'Content-Length' headers on message '")
-				errbuf.WriteString(msg.Short())
-				errbuf.WriteString(fmt.Sprintf("'; parser: %s:\n", p))
-				for _, header := range contentLengthHeaders {
-					errbuf.WriteString("\t")
-					errbuf.WriteString(header.String())
-				}
-				termErr := &sip.MalformedMessageError{
-					Err: errors.New(errbuf.String()),
-					Msg: msg.String(),
-				}
-				p.errs <- termErr
-
-				continue
-			}
-
-			contentLength = int(*(contentLengthHeaders[0].(*sip.ContentLength)))
-		} else {
-			contentLength = bodyLen
-		}
-
-		// Extract the message body.
-		p.Log().Tracef("%s reads body with length = %d bytes", p, contentLength)
-		body, err := p.input.NextChunk(contentLength)
-		if err != nil {
-			termErr := &sip.BrokenMessageError{
-				Err: fmt.Errorf("read message body failed: %w", err),
-				Msg: msg.String(),
-			}
-			p.errs <- termErr
-
-			continue
-		}
-		// RFC 3261 - 18.3.
-		if len(body) != contentLength {
-			termErr := &sip.BrokenMessageError{
-				Err: fmt.Errorf("incomplete message body: read %d bytes, expected %d bytes", len(body), contentLength),
-				Msg: msg.String(),
-			}
-			p.errs <- termErr
-
-			continue
-		}
-
-		if strings.TrimSpace(body) != "" {
-			msg.SetBody(body, false)
-		}
-
-		p.output <- msg
-	}
-	return
-}
-
-// Implements ParserFactory.SetHeaderParser.
-func (p *parser) SetHeaderParser(headerName string, headerParser HeaderParser) {
-	headerName = strings.ToLower(headerName)
-	p.headerParsers[headerName] = headerParser
+	parser := NewPacketParser(logger)
+	return parser.ParseMessage(msgData)
 }
 
 // Calculate the size of a SIP message's body, given the entire contents of the message as a byte array.
@@ -570,7 +140,7 @@ func isResponse(startLine string) bool {
 	}
 }
 
-// Parse the first line of a SIP request, e.g:
+// ParseRequestLine parse the first line of a SIP request, e.g:
 //
 //	INVITE bob@example.com SIP/2.0
 //	REGISTER jane@telco.com SIP/1.0
@@ -594,7 +164,7 @@ func ParseRequestLine(requestLine string) (
 	return
 }
 
-// Parse the first line of a SIP response, e.g:
+// ParseStatusLine parse the first line of a SIP response, e.g:
 //
 //	SIP/2.0 200 OK
 //	SIP/1.0 403 Forbidden
@@ -614,7 +184,7 @@ func ParseStatusLine(statusLine string) (
 	return
 }
 
-// parseUri converts a string representation of a URI into a Uri object.
+// ParseUri converts a string representation of a URI into a Uri object.
 // If the URI is malformed, or the URI schema is not recognised, an error is returned.
 // URIs have the general form of schema:address.
 func ParseUri(uriStr string) (uri sip.Uri, err error) {
@@ -773,7 +343,7 @@ func ParseSipUri(uriStr string) (uri sip.SipUri, err error) {
 	return
 }
 
-// Parse a text representation of a host[:port] pair.
+// ParseHostPort parse a text representation of a host[:port] pair.
 // The port may or may not be present, so we represent it with a *uint16,
 // and return 'nil' if no port was present.
 func ParseHostPort(rawText string) (host string, port *sip.Port, err error) {
@@ -832,7 +402,7 @@ func ParseHostPort(rawText string) (host string, port *sip.Port, err error) {
 	return
 }
 
-// General utility method for parsing 'key=value' parameters.
+// ParseParams is a general utility method for parsing 'key=value' parameters.
 // Takes a string (source), ensures that it begins with the 'start' character provided,
 // and then parses successive key/value pairs separated with 'sep',
 // until either 'end' is reached or there are no characters remaining.
@@ -1010,42 +580,7 @@ parseLoop:
 	return
 }
 
-// Parse a header string, producing one or more Header objects.
-// (SIP messages containing multiple headers of the same type can express them as a
-// single header containing a comma-separated argument list).
-func (p *parser) ParseHeader(headerText string) (headers []sip.Header, err error) {
-	p.Log().Tracef("parsing header \"%s\"", headerText)
-
-	headers = make([]sip.Header, 0)
-
-	colonIdx := strings.Index(headerText, ":")
-	if colonIdx == -1 {
-		err = fmt.Errorf("field name with no value in header: %s", headerText)
-		return
-	}
-
-	fieldName := strings.TrimSpace(headerText[:colonIdx])
-	lowerFieldName := strings.ToLower(fieldName)
-	fieldText := strings.TrimSpace(headerText[colonIdx+1:])
-	if headerParser, ok := p.headerParsers[lowerFieldName]; ok {
-		// We have a registered parser for this header type - use it.
-		headers, err = headerParser(lowerFieldName, fieldText)
-	} else {
-		// We have no registered parser for this header type,
-		// so we encapsulate the header data in a GenericHeader struct.
-		p.Log().Tracef("no parser for header type %s", fieldName)
-
-		header := sip.GenericHeader{
-			HeaderName: fieldName,
-			Contents:   fieldText,
-		}
-		headers = []sip.Header{&header}
-	}
-
-	return
-}
-
-// Parse a To, From or Contact header line, producing one or more logical SipHeaders.
+// Parse a "To", "From" or "Contact" header line, producing one or more logical SipHeaders.
 func parseAddressHeader(headerName string, headerText string) (
 	headers []sip.Header, err error) {
 	switch headerName {
@@ -1488,7 +1023,7 @@ func ParseAddressValue(addressText string) (
 		// be a display name.
 		if addressText[0] == '"' {
 			// The display name is within quotations.
-			// So it is comprised of all text until the closing quote.
+			// So it is composed of all text until the closing quote.
 			addressText = addressText[1:]
 			nextQuote := strings.Index(addressText, "\"")
 
@@ -1503,7 +1038,7 @@ func ParseAddressValue(addressText string) (
 			displayName = sip.String{Str: nameField}
 			addressText = addressText[nextQuote+1:]
 		} else {
-			// The display name is unquoted, so it is comprised of
+			// The display name is unquoted, so it is composed of
 			// all text until the opening angle bracket, except surrounding whitespace.
 			// According to the ABNF grammar: display-name   =  *(token LWS)/ quoted-string
 			// there are certain characters the display name cannot contain unless it's quoted,
@@ -1588,7 +1123,7 @@ func parseRecordRouteHeader(headerName string, headerText string) (headers []sip
 	return []sip.Header{&routeHeader}, nil
 }
 
-// Extract the next logical header line from the message.
+// GetNextHeaderLine extract the next logical header line from the message.
 // This may run over several actual lines; lines that start with whitespace are
 // a continuation of the previous line.
 // Therefore also return how many lines we consumed so the parent parser can
@@ -1636,7 +1171,7 @@ func findUnescaped(text string, target uint8, delims ...delimiter) int {
 	return findAnyUnescaped(text, string(target), delims...)
 }
 
-// Find the first instance of any of the targets in the given text that are not enclosed in any delimiters
+// Find the first instance of the targets in the given text that are not enclosed in any delimiters
 // from the list provided.
 func findAnyUnescaped(text string, targets string, delims ...delimiter) int {
 	escaped := false
@@ -1663,7 +1198,7 @@ func findAnyUnescaped(text string, targets string, delims ...delimiter) int {
 	return -1
 }
 
-// Splits the given string into sections, separated by one or more characters
+// SplitByWhitespace splits the given string into sections, separated by one or more characters
 // from c_ABNF_WS.
 func SplitByWhitespace(text string) []string {
 	var buffer bytes.Buffer
